fix firing rate and phase coherence measurements
Before I was using the amount of senders in spike_events
(set(spike_events['senders'])) for "counting" neurons, but this is not
correct. Imagine, for example, that some neurons have zero spikes, then
those neurons are not present in spike_events[senders] and,
consequently, are not considered. Now im using pops_dict to be sure that
all neurons are considered